requirements.txt
aiohttp~=3.11.11
setuptools~=75.6.0
discord.py~=2.4.0
psycopg[binary]~=3.2.3
psycopg[pool]~=3.2.3
Pillow==11.1.0
numpy==2.2.1
